refactor(task-board): extract ConnectGitHubDialog into its own file
index.tsx is 2705 lines; ConnectGitHubDialog is a fully self-contained
dialog that only reads its own props, so it's a clean, byte-identical
move into connect-github-dialog.tsx.
